schema_detector/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-secret-manager==2.*
requests==2.*